RANDR - Add mateconf key for disabling boot time configuration
In a lot of situations it is undesirable to have g-s-d change the XRandR
settings, because it overrides X.org customizations, leads to unnecessary mode
switches, or increases boot time.

Add a mateconf key "use_xorg_monitor_settings" to disable
apply_default_boot_configuration(), in which case the XRandR configuration will
not be touched unless there is a global or per-user configuration file.

Bug: https://bugzilla.gnome.org/show_bug.cgi?id=631388
Bug-Ubuntu: https://launchpad.net/bugs/640807