(randr) bnc#647304 - If the stored configuration fails at startup, use the fallback configurations
Previously, if a stored configuration existed but it could not be applied due
to nonmatching monitors, we would do nothing else - potentially leaving the user
in whatever startup state was used by the X server.  Now, in that condition
we simply fall back to trying the system-global configuration or the boot-time
configuration.

Signed-off-by: Federico Mena Quintero <federico@novell.com>